Implementado POST GET E DELETE usuarios

+++ marketplace-api/src/main/java/com/viborotto/marketplaceapi/models/Usuario.java
@@ -10,19 +10,20 @@ public class Usuario {
 
     @Id
     @GeneratedValue(strategy= GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     private String nome;
+
     private String email;
 
     public Usuario() {
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
